move app code into own module
